# packages/core-schemas/openapi.yaml
openapi: 3.1.0
info:
  title: DreamRight API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamRight Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}/characters:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Characters]
      summary: List all characters
      operationId: listCharacters
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of characters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCharacterList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Characters]
      summary: Create a character
      operationId: createCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterRequest'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/characters/{characterId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/characterId'

    get:
      tags: [Characters]
      summary: Get character details
      operationId: getCharacter
      responses:
        '200':
          description: Character details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Characters]
      summary: Update character
      operationId: updateCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCharacterRequest'
      responses:
        '200':
          description: Character updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Characters]
      summary: Delete character
      operationId: deleteCharacter
      responses:
        '204':
          description: Character deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

  # Character assets - Resource-oriented: POST creates the asset
  /projects/{projectId}/characters/{characterId}/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/characterId'

    get:
      tags: [Characters]
      summary: Get character assets metadata
      operationId: getCharacterAssets
      responses:
        '200':
          description: Character assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterAssets'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Characters]
      summary: Generate character portrait
      description: |
        Generates a character portrait image using AI.
        The portrait is used as a reference for panel generation.
        This is an async operation.
      operationId: createCharacterAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Asset generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Asset already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExistsError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Characters]
      summary: Regenerate character portrait
      description: Force regenerate the character portrait, replacing existing asset.
      operationId: replaceCharacterAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Asset regeneration started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/characters/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'

    post:
      tags: [Characters]
      summary: Generate all missing character assets
      description: Batch generate portraits for all characters without assets.
      operationId: createAllCharacterAssets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Batch generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

  # Location endpoints
...
  /jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/jobId'

    get:
      tags: [Jobs]
      summary: Get job status
      description: Poll for the status of an async job
      operationId: getJob
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Jobs]
      summary: Cancel job
      description: Attempt to cancel a running job
      operationId: cancelJob
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job cannot be cancelled (already completed or failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
    characterId:
      name: characterId
      in: path
      required: true
      description: Character ID
      schema:
        type: string
...
    jobId:
      name: jobId
      in: path
      required: true
      description: Job ID for async operation
      schema:
        type: string
        format: uuid
    limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    CharacterRole:
      type: string
      enum: [protagonist, antagonist, supporting, minor]
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ErrorCode:
      type: string
      enum:
        - bad_request
        - validation_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - dependency_not_met
        - asset_exists
        - rate_limited
        - internal_error
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AssetExistsError:
      type: object
      required: [code, message, asset_path]
      properties:
        code:
          type: string
          enum: [asset_exists]
        message:
          type: string
        asset_path:
          type: string
          description: Path to existing asset
        use_put_to_replace:
          type: boolean
          default: true

    # Pagination wrapper
    PaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Paginated lists
...
    PaginatedCharacterList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Job:
      type: object
      required: [id, status, type, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        type:
          type: string
          description: Type of job (story_expansion, character_asset, panel_generation, etc.)
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Progress from 0.0 to 1.0
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          nullable: true
          description: Result data when job completes
        error:
          $ref: '#/components/schemas/Error'
          nullable: true
    JobResponse:
      type: object
      required: [job_id, status, poll_url]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
          description: URL to poll for job status
    BatchJobResponse:
      type: object
      required: [job_id, status, poll_url, total_items]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
        total_items:
          type: integer
          description: Number of items to process

    # Project schemas
...
    CharacterDescription:
      type: object
      properties:
        physical:
          type: string
        personality:
          type: string
        background:
          type: string
        motivation:
          type: string
    CharacterAssets:
      type: object
      properties:
        portrait:
          type: string
          nullable: true
          description: Path to portrait image
        portrait_url:
          type: string
          format: uri
          nullable: true
          description: URL to download portrait
        three_view:
          type: object
          properties:
            front:
              type: string
              nullable: true
            side:
              type: string
              nullable: true
            back:
              type: string
              nullable: true
    Character:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/CharacterAssets'
    CreateCharacterRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
    UpdateCharacterRequest:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
    CreateCharacterAssetRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation
        portrait_only:
          type: boolean
          default: true
          description: Generate only portrait (not three-view)

    # Location schemas
...